requirements.txt
PyQt6>=6.5.0
opencv-python>=4.8.0
Pillow>=10.0.0
jsonschema>=4.17.0
numpy>=1.24.0
nuitka
zstandard
ordered-set
